Unity: #if update for building against ns2
PermissionSet is not in NS2.  We no longer need the NET_CORE define for il2cpp since that was related to building for ns16.  So just change it to a new NS2 define that il2cpp cecil building code will use the exclude this would problematic class when building our ns2 version of cecil.